requirements.txt
streamlit
python-dotenv
pymupdf
langchain-text-splitters
sentence-transformers
pinecone
groq
torch
transformers